remove old fns and refs. Move unit tests
