+++ apollo-router/src/plugins/mod.rs
@@ -47,4 +47,3 @@ pub(crate) mod telemetry;
 #[cfg(test)]
 pub(crate) mod test;
 pub(crate) mod traffic_shaping;
-pub(crate) mod error_counting_metrics;
