+++ apollo-router/src/plugins/telemetry/error_counter.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 use std::sync::Arc;
-use serde_json_bytes::Value;
+use serde_json_bytes::{Value};
 use crate::apollo_studio_interop::UsageReporting;
 use crate::context::{OPERATION_KIND, OPERATION_NAME};
 use crate::{graphql, Context};
@@ -11,7 +11,7 @@ use crate::services::{SupergraphResponse};
 use crate::services::router::ClientRequestAccepts;
 use crate::spec::query::EXTENSIONS_VALUE_COMPLETION_KEY;
 
-// TODO migrate subgraph and extended errors config
+// TODO call this for subgraph service (pre redaction), supergraph service, and _MAYBE_ router service (service unavail and invalid headers)
 pub(crate) async fn count_errors(mut response: SupergraphResponse, errors_config: &ErrorsConfiguration) -> SupergraphResponse {
     let context = response.context.clone();
     // TODO do we really need this?
@@ -27,10 +27,13 @@ pub(crate) async fn count_errors(mut response: SupergraphResponse, errors_config
 
 
     if let Some(gql_response) = response.next_response().await {
+        // TODO make mapping to add to response context to avoid double counting
+
         if !gql_response.has_next.unwrap_or(false)
             && !gql_response.subscribed.unwrap_or(false)
             && (accepts_json || accepts_wildcard)
         {
+            // TODO ensure free plan is captured
             if !gql_response.errors.is_empty() {
                 count_operation_errors(&gql_response.errors, &context, &errors_config);
             }
@@ -91,7 +94,7 @@ fn count_value_completion_errors(
             .iter()
             .filter_map(graphql::Error::from_value_completion_value)
             .collect();
-        crate::metrics::count_operation_errors(&errors, context, errors_config);
+        count_operation_errors(&errors, context, errors_config);
     }
 }
 
@@ -209,5 +212,394 @@ fn count_graphql_error(count: u64, code: Option<&str>) {
                 code = code.to_string()
             );
         }
+    }
+}
+
+
+#[cfg(test)]
+mod test {
+    use serde_json_bytes::Value;
+    use serde_json_bytes::json;
+
+    use crate::Context;
+    use crate::context::OPERATION_KIND;
+    use crate::context::OPERATION_NAME;
+    use crate::graphql;
+    use crate::json_ext::Path;
+    use crate::metrics::FutureMetricsExt;
+    use crate::plugins::telemetry::CLIENT_NAME;
+    use crate::plugins::telemetry::CLIENT_VERSION;
+    use crate::plugins::telemetry::apollo::ErrorsConfiguration;
+    use crate::plugins::telemetry::apollo::ExtendedErrorMetricsMode;
+    use crate::plugins::telemetry::error_counter::{count_operation_error_codes, count_operation_errors};
+    use crate::query_planner::APOLLO_OPERATION_ID;
+
+    #[tokio::test]
+    async fn test_count_operation_error_codes_with_extended_config_enabled() {
+        async {
+            let config = ErrorsConfiguration {
+                preview_extended_error_metrics: ExtendedErrorMetricsMode::Enabled,
+                ..Default::default()
+            };
+
+            let context = Context::default();
+            let _ = context.insert(APOLLO_OPERATION_ID, "some-id".to_string());
+            let _ = context.insert(OPERATION_NAME, "SomeOperation".to_string());
+            let _ = context.insert(OPERATION_KIND, "query".to_string());
+            let _ = context.insert(CLIENT_NAME, "client-1".to_string());
+            let _ = context.insert(CLIENT_VERSION, "version-1".to_string());
+
+            count_operation_error_codes(
+                &["GRAPHQL_VALIDATION_FAILED", "MY_CUSTOM_ERROR", "400"],
+                &context,
+                &config,
+            );
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    1,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "GRAPHQL_VALIDATION_FAILED",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "",
+                    "apollo.router.error.service" = ""
+                );
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    1,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "MY_CUSTOM_ERROR",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "",
+                    "apollo.router.error.service" = ""
+                );
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    1,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "400",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "",
+                    "apollo.router.error.service" = ""
+                );
+
+            assert_counter!(
+                    "apollo.router.graphql_error",
+                    1,
+                    code = "GRAPHQL_VALIDATION_FAILED"
+                );
+            assert_counter!("apollo.router.graphql_error", 1, code = "MY_CUSTOM_ERROR");
+            assert_counter!("apollo.router.graphql_error", 1, code = "400");
+        }
+            .with_metrics()
+            .await;
+    }
+
+    #[tokio::test]
+    async fn test_count_operation_error_codes_with_extended_config_disabled() {
+        async {
+            let config = ErrorsConfiguration {
+                preview_extended_error_metrics: ExtendedErrorMetricsMode::Disabled,
+                ..Default::default()
+            };
+
+            let context = Context::default();
+            count_operation_error_codes(
+                &["GRAPHQL_VALIDATION_FAILED", "MY_CUSTOM_ERROR", "400"],
+                &context,
+                &config,
+            );
+
+            assert_counter_not_exists!(
+                    "apollo.router.operations.error",
+                    u64,
+                    "apollo.operation.id" = "",
+                    "graphql.operation.name" = "",
+                    "graphql.operation.type" = "",
+                    "apollo.client.name" = "",
+                    "apollo.client.version" = "",
+                    "graphql.error.extensions.code" = "GRAPHQL_VALIDATION_FAILED",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "",
+                    "apollo.router.error.service" = ""
+                );
+            assert_counter_not_exists!(
+                    "apollo.router.operations.error",
+                    u64,
+                    "apollo.operation.id" = "",
+                    "graphql.operation.name" = "",
+                    "graphql.operation.type" = "",
+                    "apollo.client.name" = "",
+                    "apollo.client.version" = "",
+                    "graphql.error.extensions.code" = "MY_CUSTOM_ERROR",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "",
+                    "apollo.router.error.service" = ""
+                );
+            assert_counter_not_exists!(
+                    "apollo.router.operations.error",
+                    u64,
+                    "apollo.operation.id" = "",
+                    "graphql.operation.name" = "",
+                    "graphql.operation.type" = "",
+                    "apollo.client.name" = "",
+                    "apollo.client.version" = "",
+                    "graphql.error.extensions.code" = "400",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "",
+                    "apollo.router.error.service" = ""
+                );
+
+            assert_counter!(
+                    "apollo.router.graphql_error",
+                    1,
+                    code = "GRAPHQL_VALIDATION_FAILED"
+                );
+            assert_counter!("apollo.router.graphql_error", 1, code = "MY_CUSTOM_ERROR");
+            assert_counter!("apollo.router.graphql_error", 1, code = "400");
+        }
+            .with_metrics()
+            .await;
+    }
+
+    #[tokio::test]
+    async fn test_count_operation_errors_with_extended_config_enabled() {
+        async {
+            let config = ErrorsConfiguration {
+                preview_extended_error_metrics: ExtendedErrorMetricsMode::Enabled,
+                ..Default::default()
+            };
+
+            let context = Context::default();
+            let _ = context.insert(APOLLO_OPERATION_ID, "some-id".to_string());
+            let _ = context.insert(OPERATION_NAME, "SomeOperation".to_string());
+            let _ = context.insert(OPERATION_KIND, "query".to_string());
+            let _ = context.insert(CLIENT_NAME, "client-1".to_string());
+            let _ = context.insert(CLIENT_VERSION, "version-1".to_string());
+
+            let error = graphql::Error::builder()
+                .message("some error")
+                .extension_code("SOME_ERROR_CODE")
+                .extension("service", "mySubgraph")
+                .path(Path::from("obj/field"))
+                .build();
+
+            count_operation_errors(&[error], &context, &config);
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    1,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "SOME_ERROR_CODE",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "/obj/field",
+                    "apollo.router.error.service" = "mySubgraph"
+                );
+
+            assert_counter!("apollo.router.graphql_error", 1, code = "SOME_ERROR_CODE");
+        }
+            .with_metrics()
+            .await;
+    }
+
+    #[tokio::test]
+    async fn test_count_operation_errors_with_all_json_types_and_extended_config_enabled() {
+        async {
+            let config = ErrorsConfiguration {
+                preview_extended_error_metrics: ExtendedErrorMetricsMode::Enabled,
+                ..Default::default()
+            };
+
+            let context = Context::default();
+            let _ = context.insert(APOLLO_OPERATION_ID, "some-id".to_string());
+            let _ = context.insert(OPERATION_NAME, "SomeOperation".to_string());
+            let _ = context.insert(OPERATION_KIND, "query".to_string());
+            let _ = context.insert(CLIENT_NAME, "client-1".to_string());
+            let _ = context.insert(CLIENT_VERSION, "version-1".to_string());
+
+            let codes = [
+                json!("VALID_ERROR_CODE"),
+                json!(400),
+                json!(true),
+                Value::Null,
+                json!(["code1", "code2"]),
+                json!({"inner": "myCode"}),
+            ];
+
+            let errors = codes.map(|code| {
+                graphql::Error::from_value(json!(
+                    {
+                      "message": "error occurred",
+                      "extensions": {
+                        "code": code,
+                        "service": "mySubgraph"
+                      },
+                      "path": ["obj", "field"]
+                    }
+                    ))
+                    .unwrap()
+            });
+
+            count_operation_errors(&errors, &context, &config);
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    1,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "VALID_ERROR_CODE",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "/obj/field",
+                    "apollo.router.error.service" = "mySubgraph"
+                );
+
+            assert_counter!("apollo.router.graphql_error", 1, code = "VALID_ERROR_CODE");
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    1,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "400",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "/obj/field",
+                    "apollo.router.error.service" = "mySubgraph"
+                );
+
+            assert_counter!("apollo.router.graphql_error", 1, code = "400");
+
+            // Code is ignored for null, arrays, and objects
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    1,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "true",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "/obj/field",
+                    "apollo.router.error.service" = "mySubgraph"
+                );
+
+            assert_counter!("apollo.router.graphql_error", 1, code = "true");
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    3,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "/obj/field",
+                    "apollo.router.error.service" = "mySubgraph"
+                );
+
+            assert_counter!("apollo.router.graphql_error", 3);
+        }
+            .with_metrics()
+            .await;
+    }
+
+    #[tokio::test]
+    async fn test_count_operation_errors_with_duplicate_errors_and_extended_config_enabled() {
+        async {
+            let config = ErrorsConfiguration {
+                preview_extended_error_metrics: ExtendedErrorMetricsMode::Enabled,
+                ..Default::default()
+            };
+
+            let context = Context::default();
+            let _ = context.insert(APOLLO_OPERATION_ID, "some-id".to_string());
+            let _ = context.insert(OPERATION_NAME, "SomeOperation".to_string());
+            let _ = context.insert(OPERATION_KIND, "query".to_string());
+            let _ = context.insert(CLIENT_NAME, "client-1".to_string());
+            let _ = context.insert(CLIENT_VERSION, "version-1".to_string());
+
+            let codes = [
+                json!("VALID_ERROR_CODE"),
+                Value::Null,
+                json!("VALID_ERROR_CODE"),
+                Value::Null,
+            ];
+
+            let errors = codes.map(|code| {
+                graphql::Error::from_value(json!(
+                    {
+                      "message": "error occurred",
+                      "extensions": {
+                        "code": code,
+                        "service": "mySubgraph"
+                      },
+                      "path": ["obj", "field"]
+                    }
+                    ))
+                    .unwrap()
+            });
+
+            count_operation_errors(&errors, &context, &config);
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    2,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "VALID_ERROR_CODE",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "/obj/field",
+                    "apollo.router.error.service" = "mySubgraph"
+                );
+
+            assert_counter!("apollo.router.graphql_error", 2, code = "VALID_ERROR_CODE");
+
+            assert_counter!(
+                    "apollo.router.operations.error",
+                    2,
+                    "apollo.operation.id" = "some-id",
+                    "graphql.operation.name" = "SomeOperation",
+                    "graphql.operation.type" = "query",
+                    "apollo.client.name" = "client-1",
+                    "apollo.client.version" = "version-1",
+                    "graphql.error.extensions.code" = "",
+                    "graphql.error.extensions.severity" = "ERROR",
+                    "graphql.error.path" = "/obj/field",
+                    "apollo.router.error.service" = "mySubgraph"
+                );
+
+            assert_counter!("apollo.router.graphql_error", 2);
+        }
+            .with_metrics()
+            .await;
     }
 }
